--- src/main/java/com/impulse/controller/package-info.java
@@ -1,4 +1,0 @@
-/**
- * this package is the open API service entrance with rest controllers.
- */
-package com.impulse.controller;
